Modified code to maintain group-by-resource structure
The backend tests do not include multiple parameters in the URL so that I can pass controller functions through multiple routers. As such, to get the tests to pass I had to include functions that access tables outside of the folder's resource structure. So, I formatted the code so that the e2e software uses a group-by-resource structure with additional code that allows for passing the backend tests.