Update for tests US-08
The tests pass, but attempts at cleaning up the code by adding components to break it up, caused the tests to fail. Added frontend logging to find where the issues were happening, but they caused tests to fail. Commented out logging and used clean code where possible, but, most notably in the Dashboard a solution was not found. Next step is to create a separate branch where tests and code can be altered outside of academic constraints.