Update for 2/16/2024
Still working on cleanup.

Enforced group-by-resource.

Altered reservations and table controller for updated request resources.

Started updating routes and request resources for tests and app. Going to have to redo some of the tables controller and tests, since status is in the req body, don't need seat or finish in the route.

updated api for frontend so it's requesting properly. Will have to remove seat and finish api functions and update appropriate components to call updateStatus function instead.